pyproject.toml
[tool.poetry]
name = "moz-library"
version = "0.1.0"
description = "mozkzki's book library library."
authors = ["mozkzki <mozkzki@gmail.com>"]
maintainers = ["mozkzki <mozkzki@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mozkzki/moz-library"
repository = "https://github.com/mozkzki/moz-library"
documentation = "https://github.com/mozkzki/moz-library/wiki"
keywords = ['library', 'python']
classifiers=[
    'Development Status :: 3 - Alpha',
    'License :: OSI Approved :: MIT License',
    'Topic :: Software Development :: Libraries',
    'Intended Audience :: Developers',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.0"
python-dateutil = "^2.8.1"
selenium = "3.141.0"
beautifulsoup4 = "^4.8.2"
jinja2 = "^2.11.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = "^6.2.5"
black = "^21.9b0"
pytest-cov = "^3.0.0"
coverage = "^6.0.1"
isort = "^5.9.3"
Sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
